+++ hydrus/parser/openapi_parser.py
@@ -21,6 +21,82 @@ def try_catch_replacement(block: Any, get_this: str, default: Any) -> str:
     except KeyError:
         return default
 
+def generateEntrypoint(api_doc: HydraDoc) -> None:
+    """
+    Generates Entrypoint , Base Collection and Base Resource for the documentation
+    :param api_doc: contains the Hydra Doc created
+    """
+    api_doc.add_baseCollection()
+    api_doc.add_baseResource()
+    api_doc.gen_EntryPoint()
+
+def generate_empty_object():
+    object = {
+        "class_name":"",
+        "class_definition":HydraClass,
+        "prop_definition":list(),
+        "op_definition":list(),
+        "collection": False,
+        "path":""
+    }
+    return object
+
+
+
+def check_collection(class_name, global_,schema_obj,method):
+    collection = bool
+    # get the object type from schema block
+    try:
+        type = schema_obj["type"]
+        # if object type is array it means the method is a collection
+        if type == "array":
+            collection = True
+        else:
+            # here type should be "object"
+            collection = False
+    except KeyError:
+        collection = False
+    # checks if the method is something like 'pet/{petid}'
+    if valid_endpoint(method)=="collection" and collection==False:
+        collection=True
+    object_ = generate_empty_object()
+    # checks if the method is supported by parser at the moment or not
+    # TODO see if it is required
+    flag = check_array_param(global_["doc"]["paths"][method])
+
+    if valid_endpoint(method)!="False" :
+        try :
+            # if the class has already been parsed we will update the collection var
+            if not global_[class_name]["collection"]:
+                global_[class_name]["collection"]=True
+        except KeyError:
+            # if the class has not been parsed we will insert the object
+            object_["class_name"] = class_name
+            object_["collection"]=collection
+            global_[class_name]= object_
+    return object_
+
+
+def check_array_param(paths_):
+    for method in paths_:
+        for param in paths_[method]["parameters"]:
+            try:
+                if param["type"] == "array" and method == "get":
+                    return False
+            except KeyError:
+                pass
+    return True
+
+def valid_endpoint(path):
+    # "collection" or true means valid
+    path_ = path.split('/')
+    for subPath in path_:
+        if "{" in subPath:
+            if subPath != path_[len(path_) - 1]:
+                return "False"
+            else:
+                return "Collection"
+    return "True"
 
 def get_class_name(class_location: List[str]) -> str:
     """
@@ -29,7 +105,6 @@ def get_class_name(class_location: List[str]) -> str:
     :return: name of class
     """
     return class_location[len(class_location) - 1]
-
 
 def get_data_at_location(
         class_location: List[str], doc: Dict[str, Any]) -> Dict[str, Any]:
@@ -47,41 +122,20 @@ def get_data_at_location(
     return data
 
 
-def generateEntrypoint(api_doc: HydraDoc) -> None:
-    """
-    Generates Entrypoint , Base Collection and Base Resource for the documentation
-    :param api_doc: contains the Hydra Doc created
-    """
-    api_doc.add_baseCollection()
-    api_doc.add_baseResource()
-    api_doc.gen_EntryPoint()
-
-
-def check_if_collection(schema_block: Dict[str, Any]) -> str:
-    """
-    checks if the provided schema block represents a collection or not
-    :param schema_block: child of response object where schema is defined for return variables
-    :return: collection (string)
-    """
-
-    try:
-        type = schema_block["type"]
-        print("type is " + type)
-        if type == "array":
-            collection = "true"
+def sanitise_path(path):
+    path_ = path.split('/')
+    new_path = list()
+    for subPath in path_:
+        if "{" in subPath:
+            pass
         else:
-            collection = type
-    except KeyError:
-        collection = "false"
-    return collection
+            new_path.append(subPath)
+
+    return '/'.join(new_path)
 
 
-def get_class_details(class_location: List[str],
-                      doc: Dict["str",
-                                Any],
-                      classAndClassDefinition: Dict["str",
-                                                    HydraClass],
-                      definitionSet: Set["str"]) -> None:
+
+def get_class_details(global_,data,class_name,path="") -> None:
     """
     fetches details of class and adds the class to the dict along with the classDefinition until this point
     :param classAndClassDefinition:  dict containing class and respective defined class definition
@@ -90,22 +144,25 @@ def get_class_details(class_location: List[str],
     :param doc: the whole doc
     :return: None
     """
-    class_name = get_class_name(class_location)
+    doc = global_["doc"]
+    path = sanitise_path(path)
+    print("yoyo")
+    print(path)
+    class_name = class_name
     # we simply check if the class has been defined or not
-    if class_name not in definitionSet:
+    if class_name not in global_["class_names"]:
 
-        desc = get_data_at_location(class_location, doc)
-        classDefinition = HydraClass()
+        desc = data
         try:
             classDefinition = HydraClass(
-                class_name, class_name, desc["description"], endpoint=True)
+                class_name, class_name, desc["description"], endpoint=True,path=path)
         except KeyError:
             classDefinition = HydraClass(
-                class_name, class_name, class_name, endpoint=True)
+                class_name, class_name, class_name, endpoint=True,path=path)
 
-        properties = get_data_at_location(class_location, doc)["properties"]
+        properties = data["properties"]
         try:
-            required = get_data_at_location(class_location, doc)["required"]
+            required = data["required"]
         except KeyError:
             required = set()
         for prop in properties:
@@ -114,103 +171,138 @@ def get_class_details(class_location: List[str],
             flag = False
             if prop in required and len(required) > 0:
                 flag = True
-            classDefinition.add_supported_prop(HydraClassProp("vocab:" + prop,
-                                                              prop,
-                                                              required=flag,
-                                                              read=True,
-                                                              write=True))
-        classAndClassDefinition[class_name] = classDefinition
-        definitionSet.add(class_name)
-    else:
-        return
+            # TODO copy stuff from semantic ref branch regarding prop exists in definitionset or not
+            global_[class_name]["prop_definition"].append(HydraClassProp("vocab:" + prop,
+                                                                         prop,
+                                                                         required=flag,
+                                                                         read=True,
+                                                                         write=True))
+        global_[class_name]["path"] = path
+        global_[class_name]["class_definition"] = classDefinition
+        global_["class_names"].add(class_name)
 
 
-def check_for_ref(doc: Dict["str",
-                            Any],
-                  block: Dict[str,
-                              Any],
-                  classAndClassDefinition: Dict["str",
-                                                HydraClass],
-                  definitionSet: Set["str"]) -> Tuple[str,
-                                                      str]:
-    """
-    checks the location of schema object in the given method , can be "parameter"
-    or "responses" block and checks if the method is a collection
-    :param classAndClassDefinition: dict containing class and respective defined class definition
-    :param definitionSet: set containing the names of all parsed classes
-    :param doc: whole OAS defined doc
-    :param block: the method block from doc
-    :return: class name and collection variable
-    """
+def check_for_ref(global_, path,block):
+    # will check if there is an external ref , go to that location , add the class in globals , will also verify
+    # if we can parse this method or not , if all good will return class name
     for obj in block["responses"]:
-
         try:
-            print(block["responses"][obj]["schema"])
-            collection = check_if_collection(block["responses"][obj]["schema"])
-            print("from cfr the collection is " + collection)
             try:
                 class_location = block["responses"][obj]["schema"]["$ref"].split(
                     '/')
             except KeyError:
                 class_location = block["responses"][obj]["schema"]["items"]["$ref"].split(
                     '/')
+            object_=check_collection(class_name=class_location[2],global_=global_,schema_obj=block["responses"][obj]["schema"],
+                             method=path)
+
+            if object_["class_name"]=="":
+                # cannot parse because method not supported
+                return object_["class_name"]
             get_class_details(
-                class_location,
-                doc,
-                classAndClassDefinition,
-                definitionSet)
-            return class_location[2], collection
+                global_,get_data_at_location(class_location,global_["doc"]),get_class_name(class_location),path=path)
+            return class_location[2]
         except KeyError:
-            print(block["responses"][obj])
             pass
 
     # when we would be able to take arrays as parameters we will use
     # check_if_collection here as well c
     for obj in block["parameters"]:
-        print(obj)
         try:
-            print("we are in try for paramerters")
-            class_location = obj["schema"]["$ref"].split('/')
-            print(class_location)
+            try:
+                class_location = obj["schema"]["$ref"].split('/')
+            except KeyError:
+                class_location = obj["schema"]["items"]["$ref"].split('/')
+            object_ = check_collection(class_location[2],global_,obj["schema"],path)
+            if object_["class_name"]=="":
+                # cannot parse because method not supported
+                return object_["class_name"]
             get_class_details(
-                class_location,
-                doc,
-                classAndClassDefinition,
-                definitionSet)
-            return class_location[2], False
+                global_, get_data_at_location(class_location, global_["doc"]), get_class_name(class_location),path=path)
+            return class_location[2]
         except KeyError:
             pass
+    # cannot parse because no external ref
+    return ""
 
-    print(block)
-    return "null", "none"
+
+def allow_parameter(parameter):
+    # can add rules about param processing
+    # param can be in path too , that is already handled when we declared
+    # the class as collection from the endpoint
+    params_location = ["body"]
+    print("here")
+    print(parameter)
+    if parameter["in"] not in params_location:
+        print("yo"+parameter["in"])
+        return False
+    return True
 
 
-def get_ops(param: Dict["str", Any], method: str, class_name: str,
-            classAndClassDefinition: Dict["str", HydraClass]) -> None:
-    """
-    Extracting the operations defined for the class in the 'paths' object
-    :param classAndClassDefinition: dict containing class and respective
-                                    defined class definition
-    :param param: the path block
-    :param method: the method name ["post,"put","get"]
-    :param class_name: name of class
-    """
+def type_ref_mapping(type):
+    dataType_ref_map = dict()
+    # todo add support for byte , binary , password ,double data types
+    dataType_ref_map["integer"] = "https://schema.org/Integer"
+    dataType_ref_map["string"] = "https://schema.org/Text"
+    dataType_ref_map["long"] = "http://books.xmlschemata.org/relaxng/ch19-77199.html"
+    dataType_ref_map["float"] = "https://schema.org/Float"
+    dataType_ref_map["boolean"] = "https://schema.org/Boolean"
+    dataType_ref_map["dateTime"] = "https://schema.org/DateTime"
+    dataType_ref_map["date"] = "https://schema.org/Date"
+    return dataType_ref_map[type]
+
+def get_parameters(global_, path, method, class_name):
+    param = str
+    for parameter in global_["doc"]["paths"][path][method]["parameters"]:
+        if allow_parameter(parameter):
+            try:
+                # check if class has been parsed
+                if parameter["schema"]["$ref"].split('/')[2] in global_["class_names"]:
+                    param = "vocab:" + \
+                             parameter["schema"]["$ref"].split('/')[2]
+
+                else:
+                    # if not go to that location and parse and add
+                    get_class_details(global_,get_data_at_location(parameter["schema"]["$ref"]),
+                                      parameter["schema"]["$ref"].split('/')[2],path=path)
+                    param = "vocab:" + \
+                            parameter["schema"]["$ref"].split('/')[2]
+            except KeyError:
+                type = parameter["type"]
+                if type=="array":
+                    # TODO change this after we find a way to represent array in parameter using semantics
+                    items = parameter["schema"]["items"]
+                    try:
+                        if items["$ref"].split('/')[2] in global_["class_names"]:
+                            param = "vocab"+items["$ref"].split('/')[2]
+                        else:
+                            get_class_details(global_, get_data_at_location(items["$ref"]),
+                                      items["$ref"].split('/')[2],path=path)
+                            param = "vocab"+items["$ref"].split('/')[2]
+                    except KeyError:
+                        param = type_ref_mapping(items["type"])
+                elif type=="object":
+                    print("parameter type object !!")
+                    print("converted to string")
+                    param = "string"
+                else:
+                    param = type_ref_mapping(type)
+
+        else :
+            print("cannot process the parameter")
+            print(parameter)
+            # do further tasks
+    return param
+
+
+def get_ops(global_,path,method,class_name):
     op_method = method
     op_expects = None
-    op_name = try_catch_replacement(param[method], "summary", class_name)
+    op_name = try_catch_replacement(global_["doc"]["paths"][path][method], "summary", class_name)
     op_status = list()
+    op_expects = get_parameters(global_,path,method,class_name)
     try:
-        parameters = param[method]["parameters"]
-        for parameter in parameters:
-            try:
-                op_expects = "vocab:" + \
-                    parameter["schema"]["$ref"].split('/')[2]
-            except KeyError:
-                op_expects = parameter["schema"]["type"]
-    except KeyError:
-        op_expects = None
-    try:
-        responses = param[method]["responses"]
+        responses = global_["doc"]["paths"][path][method]["responses"]
         op_returns = None
         for response in responses:
             if response != 'default':
@@ -236,57 +328,22 @@ def get_ops(param: Dict["str", Any], method: str, class_name: str,
             {"statusCode": 200, "description": "Successful Operation"})
 
     print(" we are going to add an operation with name " + op_name)
-
-    classAndClassDefinition[class_name].add_supported_op(HydraClassOp(
+    print(op_name)
+    global_[class_name]["op_definition"].append(HydraClassOp(
         op_name, op_method.upper(), op_expects, op_returns, op_status))
 
 
-def get_paths(doc: Dict["str",
-                        Any],
-              classAndClassDefinition: Dict["str",
-                                            HydraClass],
-              definitionSet: Set["str"],
-              api_doc: HydraDoc) -> None:
-    """
-    To parse the methods defined in the paths object
-    And to add the parsed classes to the hydra api documentation
-    :param api_doc: HydraDoc defined for the spec
-    :param definitionSet: set containing all the classes already parsed
-    :param classAndClassDefinition: dict containing class and respective defined class definition
-    :param doc: the oas spec doc
-    """
-    paths = doc["paths"]
+
+def get_paths(global_) -> None:
+    paths = global_["doc"]["paths"]
     for path in paths:
         for method in paths[path]:
-            print("inside method " + method + "for path " + path)
-            class_name, collection = check_for_ref(
-                doc, paths[path][method], classAndClassDefinition, definitionSet)
-            print(
-                "the class name we got was " +
-                class_name +
-                "and the collection was " +
-                collection)
+            class_name = check_for_ref(global_,path,paths[path][method])
+            if class_name != "":
+            # do further processing 
+                get_ops(global_,path,method,class_name)
 
-            if collection != "none" and class_name != "null":
-                get_ops(
-                    paths[path],
-                    method,
-                    class_name,
-                    classAndClassDefinition)
-                possiblePath = path.split('/')[1]
-                possiblePath = possiblePath.replace(
-                    possiblePath[0], possiblePath[0].upper())
-                print("the path is " + possiblePath)
 
-                if collection is "true":
-                    api_doc.add_supported_class(
-                        classAndClassDefinition[class_name], collection=True)
-                else:
-                    api_doc.add_supported_class(classAndClassDefinition[class_name], collection=False)
-
-    for name in definitionSet:
-        api_doc.add_supported_class(
-            classAndClassDefinition[name], collection=False)
 
 
 
@@ -299,6 +356,9 @@ def parse(doc: Dict[str, Any]) -> str:
     classAndClassDefinition = dict()  # type: Dict[str,HydraClass]
     definitionSet = set()  # type: Set[str]
     info = try_catch_replacement(doc, "info", "")
+    global_= dict()
+    global_["class_names"]=definitionSet
+    global_["doc"] = doc
 
     if info != "":
         desc = try_catch_replacement(info, "description", "not defined")
@@ -312,7 +372,14 @@ def parse(doc: Dict[str, Any]) -> str:
     name = try_catch_replacement(doc, "basePath", "api")
     schemes = try_catch_replacement(doc, "schemes", "http")
     api_doc = HydraDoc(name, title, desc, name, schemes[0] + "://" + baseURL)
-    get_paths(doc, classAndClassDefinition, definitionSet, api_doc)
+    get_paths(global_)
+    for name in global_["class_names"]:
+        for prop in global_[name]["prop_definition"]:
+            global_[name]["class_definition"].add_supported_prop(prop)
+        for op in global_[name]["op_definition"]:
+            global_[name]["class_definition"].add_supported_op(op)
+        api_doc.add_supported_class(global_[name]["class_definition"],global_[name]["collection"],collection_path=global_[name]["path"])
+
     generateEntrypoint(api_doc)
     hydra_doc = api_doc.generate()
     dump = json.dumps(hydra_doc, indent=4, sort_keys=True)
@@ -326,7 +393,7 @@ def parse(doc: Dict[str, Any]) -> str:
 
 
 if __name__ == "__main__":
-    with open("../samples/petstore_mini.yaml", 'r') as stream:
+    with open("../samples/petstore_openapi.yaml", 'r') as stream:
         try:
             doc = yaml.load(stream)
         except yaml.YAMLError as exc:
